# src/api/openapi/v3.yaml
{
  "openapi": "3.1.0",
  "info": {"title": "EdgeWARN Unified API", "version": "3.0.0", "description": "The versioned, read-only EdgeWARN and EWMRS public API."},
  "servers": [{"url": "/"}],
  "paths": {
    "/api/v3": {"get": {"operationId": "getApiDiscovery", "responses": {"200": {"description": "Capability document"}}}},
    "/api/v3/openapi.json": {"get": {"operationId": "getOpenApi", "responses": {"200": {"description": "OpenAPI document"}}}},
    "/health/live": {"get": {"operationId": "getLiveness", "responses": {"200": {"description": "Live"}}}},
    "/health/ready": {"get": {"operationId": "getReadiness", "responses": {"200": {"description": "Ready"}, "503": {"description": "Producer roots unavailable"}}}},
    "/api/v3/cells": {"get": {"operationId": "listCells", "responses": {"200": {"description": "Cell collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/cells/{cellId}": {"get": {"operationId": "getCell", "parameters": [{"$ref": "#/components/parameters/cellId"}], "responses": {"200": {"description": "Cell"}, "404": {"$ref": "#/components/responses/Problem"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/storm-snapshots": {"get": {"operationId": "listStormSnapshots", "responses": {"200": {"description": "Snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/storm-snapshots/{timestamp}": {"get": {"operationId": "getStormSnapshot", "parameters": [{"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "Storm snapshot"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alert-snapshots": {"get": {"operationId": "listAlertSnapshots", "parameters": [{"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alert-snapshots/{timestamp}": {"get": {"operationId": "getAlertSnapshot", "parameters": [{"$ref": "#/components/parameters/timestamp"}, {"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert snapshot"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/alerts/{alertId}": {"get": {"operationId": "getAlert", "parameters": [{"$ref": "#/components/parameters/alertId"}, {"$ref": "#/components/parameters/source"}], "responses": {"200": {"description": "Alert"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/observations/metar": {"get": {"operationId": "listMetarObservationHours", "responses": {"200": {"description": "METAR observation hour collection"}}}},
    "/api/v3/observations/metar/{timestamp}": {"get": {"operationId": "getMetarObservations", "parameters": [{"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "METAR observations"}}}},
    "/api/v3/render-products": {"get": {"operationId": "listRenderProducts", "responses": {"200": {"description": "Render product collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}": {"get": {"operationId": "getRenderProduct", "parameters": [{"$ref": "#/components/parameters/productId"}], "responses": {"200": {"description": "Render product"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}/snapshots": {"get": {"operationId": "listRenderSnapshots", "parameters": [{"$ref": "#/components/parameters/productId"}], "responses": {"200": {"description": "Render snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/render-products/{productId}/snapshots/{timestamp}/data": {"get": {"operationId": "getRenderData", "parameters": [{"$ref": "#/components/parameters/productId"}, {"$ref": "#/components/parameters/timestamp"}], "responses": {"200": {"description": "One gzip-compressed float16 raster (top-to-bottom rows; NaN no-data)", "headers": {"X-EWMRS-Format-Version": {"schema": {"type": "string", "example": "2"}}, "X-Data-Type": {"schema": {"type": "string", "example": "float16"}}, "X-Channel-Count": {"schema": {"type": "string", "example": "1"}}, "X-Image-Width": {"schema": {"type": "string", "example": "7000"}}, "X-Image-Height": {"schema": {"type": "string", "example": "3500"}}}, "content": {"application/octet-stream": {}}}, "503": {"description": "Required EWMRS service is not active (SERVICE_NOT_ENABLED), or the requested render artifact is incomplete or invalid (INVALID_ARTIFACT)", "content": {"application/problem+json": {}}}}}},
    "/api/v3/radar-sites": {"get": {"operationId": "listRadarSites", "responses": {"200": {"description": "Radar site collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/radar-sites/{siteId}/availability": {"get": {"operationId": "getRadarAvailability", "parameters": [{"$ref": "#/components/parameters/siteId"}], "responses": {"200": {"description": "Radar availability"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/radar-sites/{siteId}/scans/{timestamp}/elevations/{elevation}/products/{productId}": {"get": {"operationId": "getRadarField", "parameters": [{"$ref":"#/components/parameters/siteId"}, {"$ref":"#/components/parameters/timestamp"}, {"$ref":"#/components/parameters/elevation"}, {"$ref":"#/components/parameters/radarProductId"}], "responses": {"200": {"description": "Gzip-compressed polar field", "content": {"application/gzip": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers": {"get": {"operationId": "listRapLayers", "responses": {"200": {"description": "RAP layer collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots": {"get": {"operationId": "listRapSnapshots", "parameters": [{"$ref":"#/components/parameters/layerId"}], "responses": {"200": {"description": "RAP snapshot collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots/{timestamp}/metadata": {"get": {"operationId": "getRapMetadata", "parameters": [{"$ref":"#/components/parameters/layerId"}, {"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "RAP metadata"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layers/{layerId}/snapshots/{timestamp}/data": {"get": {"operationId": "getRapData", "parameters": [{"$ref":"#/components/parameters/layerId"}, {"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "Uint16 data", "content": {"application/octet-stream": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/models/rap/layer-mappings": {"get": {"operationId": "getRapLayerMappings", "responses": {"200": {"description": "RAP mappings"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/analyses/wpc/surface": {"get": {"operationId": "listWpcSurfaceAnalyses", "responses": {"200": {"description": "WPC surface analysis collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/analyses/wpc/surface/{timestamp}": {"get": {"operationId": "getWpcSurfaceAnalysis", "parameters": [{"$ref":"#/components/parameters/timestamp"}], "responses": {"200": {"description": "Surface analysis GeoJSON", "content": {"application/geo+json": {}}}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}},
    "/api/v3/styles/colormaps": {"get": {"operationId": "listColormaps", "responses": {"200": {"description": "Colormap collection"}, "503": {"$ref": "#/components/responses/ServiceNotEnabled"}}}}
  },
  "components": {
    "parameters": {
      "cellId": {"name": "cellId", "in": "path", "required": true, "schema": {"type": "string"}},
      "alertId": {"name": "alertId", "in": "path", "required": true, "schema": {"type": "string"}},
      "productId": {"name": "productId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"}},
      "siteId": {"name": "siteId", "in": "path", "required": true, "schema": {"type": "string"}},
      "layerId": {"name": "layerId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9_.-]+$"}},
      "elevation": {"name": "elevation", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^\\d{1,3}(?:\\.\\d{1,2})?$"}},
      "radarProductId": {"name": "productId", "in": "path", "required": true, "schema": {"type": "string", "enum": ["DBZH", "VRADH", "WRADH", "PHIDP", "CCORH", "RHOHV", "ZDR"]}},
      "timestamp": {"name": "timestamp", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^\\d{8}-\\d{6}$"}},
      "source": {"name": "source", "in": "query", "required": true, "schema": {"type": "string", "enum": ["official", "edgewarn"]}},
      "x": {"name": "x", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}},
      "y": {"name": "y", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}}
    },
    "responses": {"Problem": {"description": "RFC 9457 problem detail", "content": {"application/problem+json": {}}}, "ServiceNotEnabled": {"description": "Required service is not active (SERVICE_NOT_ENABLED; carries service, state, and lastSeen extension members)", "content": {"application/problem+json": {}}}}
  }
}